+++ Genesis.NETCore/Entities/Requests/Initial/CardTransaction.cs
@@ -1,8 +1,6 @@
-﻿using System.ComponentModel.DataAnnotations;
+﻿using Genesis.NetCore.Entities.Attributes.Request.Financial.Business;
+using Genesis.NetCore.Entities.Attributes.Request.Financial.Funding;
 using System.Xml.Serialization;
-using Genesis.NetCore.Entities.Attributes.Request.Financial.Business;
-using Genesis.NetCore.Entities.Requests.Referential;
-using Genesis.NetCore.Validations;
 
 namespace Genesis.NetCore.Entities.Requests.Initial
 {
@@ -29,5 +27,11 @@ namespace Genesis.NetCore.Entities.Requests.Initial
 
         [XmlElement(ElementName = "business_attributes")]
         public BusinessAttributes BusinessAttributes { get; set; }
+
+        /// <summary>
+        /// Funding Transaction Params
+        /// </summary>
+        [XmlElement(ElementName = "funding")]
+        public FundingAttributes Funding { get; set; }
     }
 }